Update movie render from Star Wars API
