+++ src/App.css
@@ -7,18 +7,7 @@
   pointer-events: none;
 }
 
-.CardsContainer {
-  width: 80vw;
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-}
 
-.Loader {
-  display: block;
-  margin: 150px auto;
-}
 
 @media (prefers-reduced-motion: no-preference) {
   .App-logo {
@@ -38,4 +27,3 @@
 .App-link {
   color: rgb(112, 76, 182);
 }
-
