+++ src/components/Cards/Cards.css
@@ -10,3 +10,8 @@
   box-shadow: 0px 0px 39px rgba(95, 170, 255, 0.25);
   animation: 4s ease-in 0s infinite alternate move_eye;
 }
+
+.Loader {
+  display: block;
+  margin: 150px auto;
+}
